Create IPermissionChecker interface and implementation for authorization methods
Used for adding methods for checking whether requester has permission to see/modify requested entities.
Methods of this kind of authorization has been implemented for Shelter, AdoptionApplication and AdoptionContract. In these cases, requester has permission if the requested entity belongs to him/her (eg. Applier of AdoptionContract) or if requester is ShelterEmployee of the requested entity (eg. of the Shelter to which Animal of requested AdoptionContract belongs).