Change order of checking permission and getting Shelter entity in ShelterController
First, it will be checked whether SHelter with given id exists and if not, NotFoundException will be thrown.
Then, IPermissionChecker checks whether User is authorized to view/modify the requested Shelter and if not, ForbiddenException is thrown.
This way, when the Shelter with given id does not exist, no misleading error message is sent.
Other Controllers are left with checking the permission first because it might be more secure this way.